Need to implement LinkedLists

+++ board.h
@@ -7,15 +7,15 @@
 
 struct Board {
   byte multi[225];
-  unsigned long horiz_y[15];
-  unsigned long verti_y[15];
-  unsigned long diagr_y[21];
-  unsigned long diagl_y[21];
+  uint horiz_y[15];
+  uint verti_y[15];
+  uint diagr_y[21];
+  uint diagl_y[21];
 
-  unsigned long horiz_o[15];
-  unsigned long verti_o[15];
-  unsigned long diagr_o[21];
-  unsigned long diagl_o[21];
+  uint horiz_o[15];
+  uint verti_o[15];
+  uint diagr_o[21];
+  uint diagl_o[21];
 };
